models/predator_prey/requirements.txt
osmnx==1.9.3
torch_geometric==2.5.3
